requirements.txt
Django>=3.0,<4
djangorestframework>=3.11,<4
gunicorn>=20.0.4,<20.1
django-heroku>=0.3.1,<0.4